Fixed remote uploading from https or ftp(s)
Other ressources might work as well but everything which is not http(s) or
ftp(s) is getting an 'http://'-prefix, so one can simply omit that. This is the
same behaviour as before and results in an unkown error for other URI-schemes.
Can't think of an additional scheme which would be relevant to the real world.
I guess ftps is irrelevant enough for the time being. :P

see http://php.net/manual/en/wrappers.php

Note: Uploading will fail for ftps-servers which require SSL session reuse, like
vsftd does as default. I couldn't find a way on the fly to force fopen to count
that in.